[BE] refactor: 플래너 설정과 다른 서비스 간 의존성 제거
