+++ BE/src/main/java/com/newsainturtle/shadowmate/planner/service/DailyPlannerService.java
@@ -4,7 +4,10 @@ import com.newsainturtle.shadowmate.planner.dto.request.*;
 import com.newsainturtle.shadowmate.planner.dto.response.AddDailyTodoResponse;
 import com.newsainturtle.shadowmate.planner.dto.response.AddTimeTableResponse;
 import com.newsainturtle.shadowmate.planner.entity.DailyPlanner;
+import com.newsainturtle.shadowmate.planner_setting.entity.Category;
 import com.newsainturtle.shadowmate.user.entity.User;
+
+import java.util.List;
 
 public interface DailyPlannerService {
     AddDailyTodoResponse addDailyTodo(final User user, final AddDailyTodoRequest addDailyTodoRequest);
@@ -18,6 +21,9 @@ public interface DailyPlannerService {
     void removeDailyLike(final User user, final Long plannerWriterId, final RemoveDailyLikeRequest removeDailyLikeRequest);
     AddTimeTableResponse addTimeTable(final User user, final AddTimeTableRequest addTimeTableRequest);
     void removeTimeTable(final User user, final RemoveTimeTableRequest removeTimeTableRequest);
+    DailyPlanner getOrCreateDailyPlanner(final User user, final String date);
     DailyPlanner getDailyPlanner(final User user, final String date);
     void changeDailyTodoSequence(final User user, final ChangeDailyTodoSequenceRequest changeDailyTodoSequenceRequest);
+    long getRoutineCount(final Category category);
+    List<DailyPlanner> getDailyPlannerList(final User user);
 }
