+++ BE/src/main/java/com/newsainturtle/shadowmate/planner/service/MonthlyPlannerService.java
@@ -10,4 +10,5 @@ public interface MonthlyPlannerService {
     VisitorBookResponse addVisitorBook(final User visitor, final long ownerId, final AddVisitorBookRequest addVisitorBookRequest);
     void removeVisitorBook(final User visitor, final long ownerId, final RemoveVisitorBookRequest removeVisitorBookRequest);
     SearchVisitorBookResponse searchVisitorBook(final User visitor, final long ownerId, final long lastVisitorBookId);
+    void deleteUser(final User user);
 }
