+++ BE/src/main/java/com/newsainturtle/shadowmate/planner_setting/controller/PlannerSettingController.java
@@ -4,7 +4,10 @@ import com.newsainturtle.shadowmate.auth.service.AuthService;
 import com.newsainturtle.shadowmate.common.BaseResponse;
 import com.newsainturtle.shadowmate.config.auth.PrincipalDetails;
 import com.newsainturtle.shadowmate.planner_setting.dto.request.*;
+import com.newsainturtle.shadowmate.planner_setting.service.UserPlannerSettingService;
+import com.newsainturtle.shadowmate.planner_setting.service.PlannerRoutineService;
 import com.newsainturtle.shadowmate.planner_setting.service.PlannerSettingService;
+import com.newsainturtle.shadowmate.planner_setting.service.RoutineService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -19,7 +22,10 @@ import static com.newsainturtle.shadowmate.planner_setting.constant.PlannerSetti
 @RequiredArgsConstructor
 public class PlannerSettingController {
 
-    private final PlannerSettingService plannerSettingServiceImpl;
+    private final PlannerSettingService plannerSettingService;
+    private final PlannerRoutineService plannerRoutineService;
+    private final UserPlannerSettingService plannerAccessScopeService;
+    private final RoutineService routineService;
     private final AuthService authServiceImpl;
 
     @PostMapping("/{userId}/categories")
@@ -27,7 +33,7 @@ public class PlannerSettingController {
                                                     @PathVariable("userId") final Long userId,
                                                     @RequestBody @Valid final AddCategoryRequest addCategoryRequest) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        return ResponseEntity.ok(BaseResponse.from(SUCCESS_ADD_CATEGORY, plannerSettingServiceImpl.addCategory(principalDetails.getUser(), addCategoryRequest)));
+        return ResponseEntity.ok(BaseResponse.from(SUCCESS_ADD_CATEGORY, plannerSettingService.addCategory(principalDetails.getUser(), addCategoryRequest)));
     }
 
     @DeleteMapping("/{userId}/categories")
@@ -35,7 +41,7 @@ public class PlannerSettingController {
                                                        @PathVariable("userId") final Long userId,
                                                        @RequestBody @Valid final RemoveCategoryRequest removeCategoryRequest) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        plannerSettingServiceImpl.removeCategory(principalDetails.getUser(), removeCategoryRequest);
+        plannerRoutineService.removeCategory(principalDetails.getUser(), removeCategoryRequest);
         return ResponseEntity.accepted().body(BaseResponse.from(SUCCESS_REMOVE_CATEGORY));
     }
 
@@ -44,7 +50,7 @@ public class PlannerSettingController {
                                                        @PathVariable("userId") final Long userId,
                                                        @RequestBody @Valid final UpdateCategoryRequest updateCategoryRequest) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        plannerSettingServiceImpl.updateCategory(principalDetails.getUser(), updateCategoryRequest);
+        plannerSettingService.updateCategory(principalDetails.getUser(), updateCategoryRequest);
         return ResponseEntity.accepted().body(BaseResponse.from(SUCCESS_UPDATE_CATEGORY));
     }
 
@@ -52,14 +58,14 @@ public class PlannerSettingController {
     public ResponseEntity<BaseResponse> getCategoryColor(@AuthenticationPrincipal final PrincipalDetails principalDetails,
                                                          @PathVariable("userId") final Long userId) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        return ResponseEntity.ok(BaseResponse.from(SUCCESS_GET_CATEGORY_LIST, plannerSettingServiceImpl.getCategoryList(principalDetails.getUser())));
+        return ResponseEntity.ok(BaseResponse.from(SUCCESS_GET_CATEGORY_LIST, plannerSettingService.getCategoryList(principalDetails.getUser())));
     }
 
     @GetMapping("/{userId}/categories/colors")
     public ResponseEntity<BaseResponse> getCategoryColorList(@AuthenticationPrincipal final PrincipalDetails principalDetails,
                                                              @PathVariable("userId") final Long userId) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        return ResponseEntity.ok(BaseResponse.from(SUCCESS_GET_CATEGORY_COLOR_LIST, plannerSettingServiceImpl.getCategoryColorList()));
+        return ResponseEntity.ok(BaseResponse.from(SUCCESS_GET_CATEGORY_COLOR_LIST, plannerSettingService.getCategoryColorList()));
     }
 
     @PutMapping("/{userId}/access-scopes")
@@ -67,7 +73,7 @@ public class PlannerSettingController {
                                                        @PathVariable("userId") final Long userId,
                                                        @RequestBody @Valid final SetAccessScopeRequest setAccessScopeRequest) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        plannerSettingServiceImpl.setAccessScope(principalDetails.getUser(), setAccessScopeRequest);
+        plannerAccessScopeService.setAccessScope(principalDetails.getUser(), setAccessScopeRequest);
         return ResponseEntity.accepted().body(BaseResponse.from(SUCCESS_SET_PLANNER_ACCESS_SCOPE));
     }
 
@@ -76,14 +82,14 @@ public class PlannerSettingController {
                                                 @PathVariable("userId") final Long userId,
                                                 @RequestBody @Valid final AddDdayRequest addDdayRequest) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        return ResponseEntity.ok(BaseResponse.from(SUCCESS_ADD_DDAY, plannerSettingServiceImpl.addDday(principalDetails.getUser(), addDdayRequest)));
+        return ResponseEntity.ok(BaseResponse.from(SUCCESS_ADD_DDAY, plannerSettingService.addDday(principalDetails.getUser(), addDdayRequest)));
     }
 
     @GetMapping("/{userId}/d-days")
     public ResponseEntity<BaseResponse> getDdayList(@AuthenticationPrincipal final PrincipalDetails principalDetails,
                                                     @PathVariable("userId") final Long userId) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        return ResponseEntity.ok(BaseResponse.from(SUCCESS_GET_DDAY_LIST, plannerSettingServiceImpl.getDdayList(principalDetails.getUser())));
+        return ResponseEntity.ok(BaseResponse.from(SUCCESS_GET_DDAY_LIST, plannerSettingService.getDdayList(principalDetails.getUser())));
     }
 
     @DeleteMapping("/{userId}/d-days")
@@ -91,7 +97,7 @@ public class PlannerSettingController {
                                                    @PathVariable("userId") final Long userId,
                                                    @RequestBody @Valid final RemoveDdayRequest removeDdayRequest) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        plannerSettingServiceImpl.removeDday(principalDetails.getUser(), removeDdayRequest);
+        plannerSettingService.removeDday(principalDetails.getUser(), removeDdayRequest);
         return ResponseEntity.accepted().body(BaseResponse.from(SUCCESS_REMOVE_DDAY));
     }
 
@@ -100,7 +106,7 @@ public class PlannerSettingController {
                                                    @PathVariable("userId") final Long userId,
                                                    @RequestBody @Valid final UpdateDdayRequest updateDdayRequest) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        plannerSettingServiceImpl.updateDday(principalDetails.getUser(), updateDdayRequest);
+        plannerSettingService.updateDday(principalDetails.getUser(), updateDdayRequest);
         return ResponseEntity.accepted().body(BaseResponse.from(SUCCESS_UPDATE_DDAY));
     }
 
@@ -109,14 +115,14 @@ public class PlannerSettingController {
                                                    @PathVariable("userId") final Long userId,
                                                    @RequestBody @Valid final AddRoutineRequest addRoutineRequest) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        return ResponseEntity.ok(BaseResponse.from(SUCCESS_ADD_ROUTINE, plannerSettingServiceImpl.addRoutine(principalDetails.getUser(), addRoutineRequest)));
+        return ResponseEntity.ok(BaseResponse.from(SUCCESS_ADD_ROUTINE, plannerRoutineService.addRoutine(principalDetails.getUser(), addRoutineRequest)));
     }
 
     @GetMapping("/{userId}/routines")
     public ResponseEntity<BaseResponse> getRoutineList(@AuthenticationPrincipal final PrincipalDetails principalDetails,
                                                        @PathVariable("userId") final Long userId) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        return ResponseEntity.ok(BaseResponse.from(SUCCESS_GET_ROUTINE_LIST, plannerSettingServiceImpl.getRoutineList(principalDetails.getUser())));
+        return ResponseEntity.ok(BaseResponse.from(SUCCESS_GET_ROUTINE_LIST, routineService.getRoutineList(principalDetails.getUser())));
     }
 
     @DeleteMapping("/{userId}/routines")
@@ -124,7 +130,7 @@ public class PlannerSettingController {
                                                       @PathVariable("userId") final Long userId,
                                                       @RequestBody @Valid final RemoveRoutineRequest removeRoutineRequest) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        plannerSettingServiceImpl.removeRoutine(principalDetails.getUser(), removeRoutineRequest);
+        routineService.removeRoutine(principalDetails.getUser(), removeRoutineRequest);
         return ResponseEntity.ok(BaseResponse.from(SUCCESS_REMOVE_ROUTINE));
     }
 
@@ -133,7 +139,7 @@ public class PlannerSettingController {
                                                       @PathVariable("userId") final Long userId,
                                                       @RequestBody @Valid final UpdateRoutineRequest updateRoutineRequest) {
         authServiceImpl.certifyUser(userId, principalDetails.getUser());
-        plannerSettingServiceImpl.updateRoutine(principalDetails.getUser(), updateRoutineRequest);
+        plannerRoutineService.updateRoutine(principalDetails.getUser(), updateRoutineRequest);
         return ResponseEntity.ok(BaseResponse.from(SUCCESS_UPDATE_ROUTINE));
     }
 }
