+++ BE/src/main/java/com/newsainturtle/shadowmate/user/controller/UserController.java
@@ -3,6 +3,7 @@ package com.newsainturtle.shadowmate.user.controller;
 import com.newsainturtle.shadowmate.auth.service.AuthService;
 import com.newsainturtle.shadowmate.common.BaseResponse;
 import com.newsainturtle.shadowmate.config.auth.PrincipalDetails;
+import com.newsainturtle.shadowmate.planner_setting.service.UserPlannerSettingService;
 import com.newsainturtle.shadowmate.user.dto.request.UpdateIntroductionRequest;
 import com.newsainturtle.shadowmate.user.dto.request.UpdatePasswordRequest;
 import com.newsainturtle.shadowmate.user.dto.request.UpdateUserRequest;
@@ -24,6 +25,7 @@ public class UserController {
 
     private final UserService userService;
     private final AuthService authService;
+    private final UserPlannerSettingService userPlannerSettingService;
 
     @GetMapping("/{userId}/profiles")
     public ResponseEntity<BaseResponse> getProfile(@PathVariable final Long userId) {
@@ -34,7 +36,8 @@ public class UserController {
 
     @GetMapping("/{userId}/introduction")
     public ResponseEntity<BaseResponse> searchIntroduction(@PathVariable("userId") final Long userId) {
-        return ResponseEntity.ok(BaseResponse.from(SUCCESS_SEARCH_INTRODUCTION, userService.searchIntroduction(userId)));
+        return ResponseEntity
+                .ok(BaseResponse.from(SUCCESS_SEARCH_INTRODUCTION, userService.searchIntroduction(userId)));
     }
 
     @PutMapping("/{userId}/mypages")
@@ -59,14 +62,15 @@ public class UserController {
     public ResponseEntity<BaseResponse> deleteUser(@AuthenticationPrincipal final PrincipalDetails principalDetails,
                                                    @PathVariable("userId") final Long userId) {
         authService.certifyUser(userId, principalDetails.getUser());
-        userService.deleteUser(principalDetails.getUser());
+        userPlannerSettingService.deleteUser(principalDetails.getUser());
         return ResponseEntity.ok(BaseResponse.from(SUCCESS_DELETE_USER));
     }
 
     @PutMapping("/{userId}/introduction")
-    public ResponseEntity<BaseResponse> updateIntroduction(@AuthenticationPrincipal final PrincipalDetails principalDetails,
-                                                           @PathVariable("userId") final Long userId,
-                                                           @RequestBody @Valid final UpdateIntroductionRequest updateIntroductionRequest) {
+    public ResponseEntity<BaseResponse> updateIntroduction(
+            @AuthenticationPrincipal final PrincipalDetails principalDetails,
+            @PathVariable("userId") final Long userId,
+            @RequestBody @Valid final UpdateIntroductionRequest updateIntroductionRequest) {
         authService.certifyUser(userId, principalDetails.getUser());
         userService.updateIntroduction(userId, updateIntroductionRequest);
         return ResponseEntity.ok(BaseResponse.from(SUCCESS_UPDATE_INTRODUCTION));
